Add the compilation directive : -Wconversion Now there is possible "auto" cast.
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/scalfmm/scalfmm/trunk@189 2616d619-271b-44dc-8df4-d4a8f33a7222

+++ Src/Containers/FTreeCoordinate.hpp
@@ -159,11 +159,11 @@ public:
         this->z = 0;
 
         for(int indexLevel = 0; indexLevel < inLevel ; ++indexLevel){
-            z |= (inIndex & mask);
+            z |= long(inIndex & mask);
             inIndex >>= 1;
-            y |= (inIndex & mask);
+            y |= long(inIndex & mask);
             inIndex >>= 1;
-            x |= (inIndex & mask);
+            x |= long(inIndex & mask);
 
             mask <<= 1;
         }
